refactor: 폴더구조 컨벤션 적용
