+++ src/app/user/create-group/page.tsx
@@ -3,9 +3,10 @@ import UserDashboardPage from "@/app/user/page";
 import Modal from "@/common/component/Modal";
 import ToastProvider from "@/common/component/Toast";
 import { useToast } from "@/common/hook/useToast";
-import CreateGroupForm from "@/view/user/create-group/component/CreateGroupForm";
-import LinkWithCopy from "@/view/user/create-group/component/LinkWithCopy";
-import { wrapper } from "@/view/user/create-group/component/index.css";
+import CreateGroupForm from "@/view/user/create-group/CreateGroupForm";
+
+import LinkWithCopy from "@/view/user/create-group/LinkWithCopy";
+import { wrapper } from "@/view/user/create-group/index.css";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
