--- src/view/login/component/Heading/index.tsx
@@ -1,7 +1,0 @@
-import { headingStyle } from "./index.css";
-
-const Heading = () => {
-  return <h2 className={headingStyle}>알고허브로 로그인</h2>;
-};
-
-export default Heading;
