+++ src/view/user/create-group/LinkWithCopy/index.tsx
